Implemented Consignee with a debugger

+++ ticket.cpp
@@ -10,7 +10,7 @@ using namespace std;
 hash<string> strHash;
 Ticket::Ticket() {
   string notHashed = "";
-  for(int i=0; i < 5; i++)
+  for(size_t i=0; i < 5; i++)
     notHashed += notHashed + to_string(rand());
 
   size_t hashed = strHash(notHashed);
